OSS prep: fix all warnings, cross-platform CI, notarized releases, docs
- Fix -Wimplicit-int-float-conversion and -Wshadow-field warnings; our
  sources now build clean under juce_recommended_warning_flags
- Portable tests: std::filesystem instead of mkdir -p, no M_PI (MSVC)
- CI matrix builds + runs dsp_tests on macOS/Windows/Linux
- Release workflow on v* tags: universal macOS binary, codesign +
  notarytool + staple via secrets, zips for all three platforms
- scripts/notarize.sh for local notarization
- Presets no longer overwrite the Mix parameter (perceived reset bug)
- README rewrite for Chorale, CONTRIBUTING.md, refreshed demos/

+++ src/dsp/YinTracker.cpp
@@ -36,7 +36,7 @@ PitchEstimate YinTracker::analyze (const float* x)
     for (int tau = 1; tau <= tauMax; ++tau)
     {
         running += d[(size_t) tau];
-        cmnd[(size_t) tau] = running > 0.0 ? (float) (d[(size_t) tau] * tau / running) : 1.0f;
+        cmnd[(size_t) tau] = running > 0.0 ? (float) (d[(size_t) tau] * (double) tau / running) : 1.0f;
     }
 
     int best = -1;
